oracle_v2/requirements.txt
web3==6.10.0
python-dotenv==1.0.1
requests==2.31.0
ipfshttpclient==0.8.0a2
pandas==2.2.2
schedule==1.2.1
loguru==0.7.2
jsonschema==4.22.0
